tmux: Add bindings for window creation
M-c for new window in current path, C-c to change session's current
path, c for new window in session's default path. Originally wanted M-c
to be C binding, but that's bound to "Customize options." Not sure if
it's worth it to keep this binding with how little I use it.

Switched current bindings for sane window splitting ("|" for vertical and
"-" for horizontal) to use current directory by default. Use Alt
modifier to split with session's default directory. Also documented
alternative Shift modifier.